requirements.txt
Pillow
boto
base58
django